add implementation of unit types
This commit adds a new file unit.py which defines the type Unit and some subclasses. The idea is that a unit's type (land, naval, etc.) can be identified by which subclass it is. The top-level Unit class implements __hash__(), which takes the unit type into account in the hashing.